+++ Packages/com.github.kurotu.vrc-quest-tools/Editor/Inspector/ToonStandardConvertSettingsDrawer.cs
@@ -72,6 +72,11 @@ namespace KRT.VRCQuestTools.Inspector
                 fieldRect.y += EditorGUIUtility.standardVerticalSpacing;
             }
 
+            var generateShadowRamp = property.FindPropertyRelative("generateShadowRamp");
+            EditorGUI.PropertyField(fieldRect, generateShadowRamp, new GUIContent(i18n.ToonStandardConvertSettingsGenerateShadowRampLabel));
+            fieldRect.y += EditorGUIUtility.singleLineHeight;
+            fieldRect.y += EditorGUIUtility.standardVerticalSpacing;
+
             var fallbackShadowRamp = property.FindPropertyRelative("fallbackShadowRamp");
             using (var scope = new EditorGUI.PropertyScope(fieldRect, new GUIContent(i18n.ToonStandardConvertSettingsFallbackShadowRampLabel), fallbackShadowRamp))
             {
@@ -100,6 +105,65 @@ namespace KRT.VRCQuestTools.Inspector
                 }
             }
 
+            EditorGUI.LabelField(fieldRect, new GUIContent(i18n.ToonStandardConvertSettingsFeaturesLabel));
+
+            var buttonRect = fieldRect;
+            buttonRect.x = fieldRect.x + EditorGUIUtility.labelWidth;
+            buttonRect.width = (fieldRect.width - EditorGUIUtility.labelWidth) / 2;
+            if (GUI.Button(buttonRect, new GUIContent(i18n.SelectAllButtonLabel)))
+            {
+                if (property.managedReferenceValue is ToonStandardConvertSettings settings)
+                {
+                    settings.SetAllFeatures(true);
+                    property.serializedObject.ApplyModifiedProperties();
+                }
+            }
+            buttonRect.x += buttonRect.width;
+            if (GUI.Button(buttonRect, new GUIContent(i18n.DeselectAllButtonLabel)))
+            {
+                if (property.managedReferenceValue is ToonStandardConvertSettings settings)
+                {
+                    settings.SetAllFeatures(false);
+                    property.serializedObject.ApplyModifiedProperties();
+                }
+            }
+
+            fieldRect.y += EditorGUIUtility.singleLineHeight;
+            fieldRect.y += EditorGUIUtility.standardVerticalSpacing;
+
+            using (new EditorGUI.IndentLevelScope())
+            {
+                var useNormalMap = property.FindPropertyRelative("useNormalMap");
+                EditorGUI.PropertyField(fieldRect, useNormalMap, new GUIContent(i18n.ToonStandardConvertSettingsFeaturesNormalMapLabel));
+                fieldRect.y += EditorGUIUtility.singleLineHeight;
+                fieldRect.y += EditorGUIUtility.standardVerticalSpacing;
+
+                var useEmission = property.FindPropertyRelative("useEmission");
+                EditorGUI.PropertyField(fieldRect, useEmission, new GUIContent(i18n.ToonStandardConvertSettingsFeaturesEmissionLabel));
+                fieldRect.y += EditorGUIUtility.singleLineHeight;
+                fieldRect.y += EditorGUIUtility.standardVerticalSpacing;
+
+                var useOcclusion = property.FindPropertyRelative("useOcclusion");
+                EditorGUI.PropertyField(fieldRect, useOcclusion, new GUIContent(i18n.ToonStandardConvertSettingsFeaturesOcclusionLabel));
+                fieldRect.y += EditorGUIUtility.singleLineHeight;
+                fieldRect.y += EditorGUIUtility.standardVerticalSpacing;
+
+                var useSpecular = property.FindPropertyRelative("useSpecular");
+                EditorGUI.PropertyField(fieldRect, useSpecular, new GUIContent(i18n.ToonStandardConvertSettingsFeaturesSpecularLabel));
+                fieldRect.y += EditorGUIUtility.singleLineHeight;
+                fieldRect.y += EditorGUIUtility.standardVerticalSpacing;
+
+                var useMatcap = property.FindPropertyRelative("useMatcap");
+                EditorGUI.PropertyField(fieldRect, useMatcap, new GUIContent(i18n.ToonStandardConvertSettingsFeaturesMatcapLabel));
+                fieldRect.y += EditorGUIUtility.singleLineHeight;
+                fieldRect.y += EditorGUIUtility.standardVerticalSpacing;
+
+                var useRim = property.FindPropertyRelative("useRimLighting");
+                EditorGUI.PropertyField(fieldRect, useRim, new GUIContent(i18n.ToonStandardConvertSettingsFeaturesRimLightingLabel));
+                fieldRect.y += EditorGUIUtility.singleLineHeight;
+                fieldRect.y += EditorGUIUtility.standardVerticalSpacing;
+            }
+
             EditorGUI.indentLevel--;
 
             return fieldRect;
@@ -116,6 +180,8 @@ namespace KRT.VRCQuestTools.Inspector
             height += EditorGUI.GetPropertyHeight(property.FindPropertyRelative("mobileTextureFormat"));
             height += EditorGUIUtility.standardVerticalSpacing;
 
+            height += EditorGUI.GetPropertyHeight(property.FindPropertyRelative("generateShadowRamp"));
+            height += EditorGUIUtility.standardVerticalSpacing;
             var fallbackShadowRamp = property.FindPropertyRelative("fallbackShadowRamp");
             height += EditorGUIUtility.singleLineHeight;
             height += EditorGUIUtility.standardVerticalSpacing;
@@ -124,6 +190,22 @@ namespace KRT.VRCQuestTools.Inspector
                 height += EditorGUI.GetPropertyHeight(fallbackShadowRamp);
                 height += EditorGUIUtility.standardVerticalSpacing;
             }
+
+            height += EditorGUIUtility.singleLineHeight; // Features label
+            height += EditorGUIUtility.standardVerticalSpacing;
+            height += EditorGUI.GetPropertyHeight(property.FindPropertyRelative("useNormalMap"));
+            height += EditorGUIUtility.standardVerticalSpacing;
+            height += EditorGUI.GetPropertyHeight(property.FindPropertyRelative("useEmission"));
+            height += EditorGUIUtility.standardVerticalSpacing;
+            height += EditorGUI.GetPropertyHeight(property.FindPropertyRelative("useOcclusion"));
+            height += EditorGUIUtility.standardVerticalSpacing;
+            height += EditorGUI.GetPropertyHeight(property.FindPropertyRelative("useSpecular"));
+            height += EditorGUIUtility.standardVerticalSpacing;
+            height += EditorGUI.GetPropertyHeight(property.FindPropertyRelative("useMatcap"));
+            height += EditorGUIUtility.standardVerticalSpacing;
+            height += EditorGUI.GetPropertyHeight(property.FindPropertyRelative("useRimLighting"));
+            height += EditorGUIUtility.standardVerticalSpacing;
+
             return height;
         }