+++ Packages/com.github.kurotu.vrc-quest-tools/Editor/Models/MaterialGenerators/ToonStandardGenerator.cs
@@ -16,7 +16,7 @@ namespace KRT.VRCQuestTools.Models
         /// <summary>
         /// Settings for the material conversion.
         /// </summary>
-        protected readonly ToonStandardConvertSettings settings;
+        protected readonly ToonStandardConvertSettings Settings;
 
         /// <summary>
         /// Shared black texture to disable emission.
@@ -30,7 +30,7 @@ namespace KRT.VRCQuestTools.Models
         /// <param name="sharedBlackTexture">Shared black texture to disable emission.</param>
         internal ToonStandardGenerator(ToonStandardConvertSettings settings, Texture2D sharedBlackTexture)
         {
-            this.settings = settings;
+            this.Settings = settings;
             this.sharedBlackTexture = sharedBlackTexture;
         }
 
@@ -73,14 +73,13 @@ namespace KRT.VRCQuestTools.Models
         /// <inheritdoc/>
         public AsyncCallbackRequest GenerateMaterial(MaterialBase material, UnityEditor.BuildTarget buildTarget, bool saveTextureAsPng, string texturesPath, Action<Material> completion)
         {
-#if VQT_HAS_VRCSDK_TOON_STANDARD
             if (!(material is IToonStandardConvertable))
             {
                 var toonLitConvertSettings = new ToonLitConvertSettings
                 {
-                    generateQuestTextures = settings.generateQuestTextures,
-                    maxTextureSize = settings.maxTextureSize,
-                    mobileTextureFormat = settings.mobileTextureFormat,
+                    generateQuestTextures = Settings.generateQuestTextures,
+                    maxTextureSize = Settings.maxTextureSize,
+                    mobileTextureFormat = Settings.mobileTextureFormat,
                     mainTextureBrightness = 1.0f,
                     generateShadowFromNormalMap = true,
                 };
@@ -89,13 +88,13 @@ namespace KRT.VRCQuestTools.Models
                     var newMaterial = new ToonStandardMaterialWrapper();
                     newMaterial.Name = material.Material.name;
                     newMaterial.MainTexture = newMat.mainTexture;
-                    newMaterial.ShadowRamp = settings.fallbackShadowRamp;
+                    newMaterial.ShadowRamp = Settings.fallbackShadowRamp;
                     completion?.Invoke(newMaterial);
                 });
             }
 
             ToonStandardMaterialWrapper newMaterial;
-            if (settings.generateQuestTextures)
+            if (Settings.generateQuestTextures)
             {
                 newMaterial = new ToonStandardMaterialWrapper();
                 newMaterial.Name = material.Material.name;
@@ -103,7 +102,7 @@ namespace KRT.VRCQuestTools.Models
 
                 if (GetUseMainTexture())
                 {
-                    MaterialGeneratorUtility.GenerateTexture(material.Material, settings, "main", saveTextureAsPng, texturesPath, (compl) => GenerateMainTexture(compl), (t) =>
+                    MaterialGeneratorUtility.GenerateTexture(material.Material, Settings, "main", saveTextureAsPng, texturesPath, (compl) => GenerateMainTexture(compl), (t) =>
                     {
                         newMaterial.MainTexture = t;
                         newMaterial.MainColor = new Color(1, 1, 1, 1);
@@ -114,12 +113,12 @@ namespace KRT.VRCQuestTools.Models
                     newMaterial.MainColor = GetMainColor();
                 }
 
-                if (GetUseNormalMap())
+                if (GetUseNormalMap() && Settings.useNormalMap)
                 {
                     newMaterial.UseNormalMap = true;
                     var isMobile = buildTarget == UnityEditor.BuildTarget.Android || buildTarget == UnityEditor.BuildTarget.iOS;
                     var outputRGB = saveTextureAsPng || isMobile;
-                    MaterialGeneratorUtility.GenerateNormalMap(material.Material, settings, "normal", saveTextureAsPng, texturesPath, (compl) => GenerateNormalMap(outputRGB, compl), (t) =>
+                    MaterialGeneratorUtility.GenerateNormalMap(material.Material, Settings, "normal", saveTextureAsPng, texturesPath, (compl) => GenerateNormalMap(outputRGB, compl), (t) =>
                     {
                         newMaterial.NormalMap = t;
                         (newMaterial.NormalMapTextureScale, newMaterial.NormalMapTextureOffset) = GetNormalMapST();
@@ -131,12 +130,21 @@ namespace KRT.VRCQuestTools.Models
 
                 if (GetUseShadowRamp())
                 {
-                    MaterialGeneratorUtility.GenerateTexture(material.Material, settings, "shadowRamp", saveTextureAsPng, texturesPath, (compl) => GenerateShadowRamp(compl), (t) =>
+                    if (Settings.generateShadowRamp)
                     {
-                        newMaterial.ShadowRamp = t;
+                        MaterialGeneratorUtility.GenerateTexture(material.Material, Settings, "shadowRamp", saveTextureAsPng, texturesPath, (compl) => GenerateShadowRamp(compl), (t) =>
+                        {
+                            newMaterial.ShadowRamp = t;
+                            newMaterial.ShadowBoost = 0.0f;
+                            newMaterial.ShadowTint = 0.0f;
+                        }).WaitForCompletion();
+                    }
+                    else
+                    {
+                        newMaterial.ShadowRamp = Settings.fallbackShadowRamp;
                         newMaterial.ShadowBoost = 0.0f;
-                        newMaterial.ShadowTint = 0.0f;
-                    }).WaitForCompletion();
+                        newMaterial.ShadowTint = 0.5f;
+                    }
                 }
                 else
                 {
@@ -145,10 +153,11 @@ namespace KRT.VRCQuestTools.Models
 
                 newMaterial.MinBrightness = GetMinBrightness();
 
-                if (GetUseEmission()) {
+                if (GetUseEmission() && Settings.useEmission)
+                {
                     if (GetUseEmissionMap())
                     {
-                        MaterialGeneratorUtility.GenerateTexture(material.Material, settings, "emission", saveTextureAsPng, texturesPath, (compl) => GenerateEmissionMap(compl), (t) =>
+                        MaterialGeneratorUtility.GenerateTexture(material.Material, Settings, "emission", saveTextureAsPng, texturesPath, (compl) => GenerateEmissionMap(compl), (t) =>
                         {
                             newMaterial.EmissionMap = t;
                             newMaterial.EmissionColor = new Color(1, 1, 1, 1);
@@ -156,20 +165,22 @@ namespace KRT.VRCQuestTools.Models
                     }
                     else
                     {
-                            newMaterial.EmissionColor = GetEmissionColor();
+                        newMaterial.EmissionColor = GetEmissionColor();
                     }
-                } else {
+                }
+                else
+                {
                     newMaterial.EmissionMap = sharedBlackTexture;
                     newMaterial.EmissionColor = Color.black;
                 }
 
-                if (GetUseOcclusionMap())
+                if (GetUseOcclusionMap() && Settings.useOcclusion)
                 {
                     newMaterial.UseOcclusion = true;
                     masks.Add(MaskType.OcculusionMap);
                 }
 
-                if (GetUseSpecular())
+                if (GetUseSpecular() && Settings.useSpecular)
                 {
                     newMaterial.UseSpecular = true;
                     if (GetUseMetallicMap())
@@ -188,10 +199,10 @@ namespace KRT.VRCQuestTools.Models
                     newMaterial.Reflectance = GetReflectance();
                 }
 
-                if (GetUseMatcap())
+                if (GetUseMatcap() && Settings.useMatcap)
                 {
                     newMaterial.UseMatcap = true;
-                    MaterialGeneratorUtility.GenerateTexture(material.Material, settings, "matcap", saveTextureAsPng, texturesPath, (compl) => GenerateMatcap(compl), (t) =>
+                    MaterialGeneratorUtility.GenerateTexture(material.Material, Settings, "matcap", saveTextureAsPng, texturesPath, (compl) => GenerateMatcap(compl), (t) =>
                     {
                         newMaterial.Matcap = t;
                     }).WaitForCompletion();
@@ -204,7 +215,7 @@ namespace KRT.VRCQuestTools.Models
                     newMaterial.MatcapType = GetMapcapType();
                 }
 
-                if (GetUseRimLighting())
+                if (GetUseRimLighting() && Settings.useRimLighting)
                 {
                     newMaterial.UseRimLighting = true;
                     newMaterial.RimColor = GetRimColor();
@@ -218,6 +229,7 @@ namespace KRT.VRCQuestTools.Models
                 if (masks.Count > 0)
                 {
                     var texturePacks = new List<TexturePack>();
+
                     // per 4 masks, generate a texture pack.
                     for (int i = 0; i < masks.Count; i += 4)
                     {
@@ -235,7 +247,7 @@ namespace KRT.VRCQuestTools.Models
                     foreach (var pack in texturePacks)
                     {
                         var name = $"mask_{pack.R}_{pack.G}_{pack.B}_{pack.A}";
-                        MaterialGeneratorUtility.GenerateTexture(material.Material, settings, name, saveTextureAsPng, texturesPath, (compl) => GeneratePackedMask(pack, compl), (t) =>
+                        MaterialGeneratorUtility.GenerateTexture(material.Material, Settings, name, saveTextureAsPng, texturesPath, (compl) => GeneratePackedMask(pack, compl), (t) =>
                         {
                             foreach (var mask in pack.GetMasks())
                             {
@@ -246,6 +258,7 @@ namespace KRT.VRCQuestTools.Models
                                     case MaskType.DetailMask:
                                         newMaterial.DetailMask = t;
                                         newMaterial.DetailMaskChannel = mask.Channel;
+
                                         // TODO: DetailMask ST.
                                         break;
                                     case MaskType.MetallicMap:
@@ -282,9 +295,6 @@ namespace KRT.VRCQuestTools.Models
             }
 
             return new ResultRequest<Material>(newMaterial, completion);
-#else
-            throw new InvalidOperationException("VRCSDK 3.8.1 or later is required for Toon Standard.");
-#endif
         }
 
         /// <inheritdoc/>